Add userEventsReducer and tests
