+++ src/reducers/index.js
@@ -1,16 +1,14 @@
 import { combineReducers } from 'redux';
 import { eventsReducer } from './eventsReducer/eventsReducer';
 import { userReducer } from './userReducer/userReducer';
-import { wishlistReducer } from './wishlistReducer/wishlistReducer';
-import { attendingReducer } from './attendingReducer/attendingReducer';
+import { userEventsReducer } from './userEventsReducer/userEventsReducer';
 import { errorReducer } from './errorReducer/errorReducer';
 import { loadingReducer } from './loadingReducer/loadingReducer';
 
 export const rootReducer = combineReducers({
   events: eventsReducer,
   user: userReducer,
-  wishlist: wishlistReducer,
-  attending: attendingReducer,
+  userEvents: userEventsReducer,
   loading: loadingReducer,
   error: errorReducer,
 });
